refactor: CommnetsResponse 코드 분리, Document 오타 수정
